+++ src/index.js
@@ -2,9 +2,10 @@ let knex = require('knex');
 let JSData = require('js-data');
 let map = require('mout/array/map');
 let underscore = require('mout/string/underscore');
+let unique = require('mout/array/unique');
 let toString = require('mout/lang/toString');
 let { DSUtils } = JSData;
-let { keys, isEmpty, upperCase, omit, contains, forOwn, deepMixIn, forEach, isObject, isString, removeCircular } = DSUtils;
+let { keys, isEmpty, upperCase, omit, contains, forOwn, deepMixIn, filter, forEach, isObject, isString, removeCircular } = DSUtils;
 
 let reserved = [
   'orderBy',
@@ -148,7 +149,8 @@ class DSSqlAdapter {
 
           forEach(resourceConfig.relationList, def => {
             let relationName = def.relation;
-            if (contains(options.with, relationName)) {
+            if (contains(options.with, relationName) || contains(options.with, def.localField)) {
+              DSUtils.remove(options.with, relationName);
               let task;
               let params = {};
               if (resourceConfig.allowSimpleWhere) {
@@ -163,15 +165,15 @@ class DSSqlAdapter {
               if (def.type === 'hasMany' && params[def.foreignKey]) {
                 task = this.findAll(resourceConfig.getResource(relationName), params, options);
               } else if (def.type === 'hasOne') {
-                if (def.localKey && instance[def.localKey]) {
-                  task = this.find(resourceConfig.getResource(relationName), instance[def.localKey], options);
+                if (def.localKey && DSUtils.get(instance, def.localKey)) {
+                  task = this.find(resourceConfig.getResource(relationName), DSUtils.get(instance, def.localKey), options);
                 } else if (def.foreignKey && params[def.foreignKey]) {
                   task = this.findAll(resourceConfig.getResource(relationName), params, options).then(hasOnes => {
                     return hasOnes.length ? hasOnes[0] : null;
                   });
                 }
-              } else if (instance[def.localKey]) {
-                task = this.find(resourceConfig.getResource(relationName), instance[def.localKey], options);
+              } else if (DSUtils.get(instance, def.localKey)) {
+                task = this.find(resourceConfig.getResource(relationName), DSUtils.get(instance, def.localKey), options);
               }
 
               if (task) {
@@ -185,13 +187,115 @@ class DSSqlAdapter {
         }
       })
       .then(loadedRelations => {
-        forEach(fields, (field, index) => instance[field] = loadedRelations[index]);
+        forEach(fields, (field, index) => DSUtils.set(instance, field, loadedRelations[index]));
         return instance;
       });
   }
 
   findAll(resourceConfig, params, options) {
-    return filterQuery.call(this, resourceConfig, params, options);
+    let items = null;
+    options = options || {};
+    options.with = options.with || [];
+    return filterQuery.call(this, resourceConfig, params, options).then(_items => {
+      items = _items;
+      let tasks = [];
+      forEach(resourceConfig.relationList, def => {
+        let relationName = def.relation;
+        let relationDef = resourceConfig.getResource(relationName);
+        let containedName = null;
+        if (contains(options.with, relationName)) {
+          containedName = relationName;
+        } else if (contains(options.with, def.localField)) {
+          containedName = def.localField;
+        }
+        if (containedName) {
+          let __options = DSUtils.deepMixIn({}, options.orig ? options.orig() : options);
+          __options = DSUtils._(relationDef, __options);
+          DSUtils.remove(__options.with, containedName);
+          forEach(__options.with, (relation, i) => {
+            if (relation && relation.indexOf(containedName) === 0 && relation.length >= containedName.length && relation[containedName.length] === '.') {
+              __options.with[i] = relation.substr(containedName.length + 1);
+            }
+          });
+
+          let task;
+
+          if ((def.type === 'hasOne' || def.type === 'hasMany') && def.foreignKey) {
+            task = this.findAll(resourceConfig.getResource(relationName), {
+              where: {
+                [def.foreignKey]: {
+                  'in': filter(map(items, item => DSUtils.get(item, resourceConfig.idAttribute)), x => x)
+                }
+              }
+            }, __options).then(relatedItems => {
+              forEach(items, item => {
+                let attached = [];
+                forEach(relatedItems, relatedItem => {
+                  if (DSUtils.get(relatedItem, def.foreignKey) === item[resourceConfig.idAttribute]) {
+                    attached.push(relatedItem);
+                  }
+                });
+                if (def.type === 'hasOne' && attached.length) {
+                  DSUtils.set(item, def.localField, attached[0]);
+                } else {
+                  DSUtils.set(item, def.localField, attached);
+                }
+              });
+              return relatedItems;
+            });
+          } else if (def.type === 'hasMany' && def.localKeys) {
+            let localKeys = [];
+            forEach(items, item => {
+              let itemKeys = item[def.localKeys] || [];
+              itemKeys = Array.isArray(itemKeys) ? itemKeys : keys(itemKeys);
+              localKeys = localKeys.concat(itemKeys || []);
+            });
+            task = this.findAll(resourceConfig.getResource(relationName), {
+              where: {
+                [relationDef.idAttribute]: {
+                  'in': filter(unique(localKeys), x => x)
+                }
+              }
+            }, __options).then(relatedItems => {
+              forEach(items, item => {
+                let attached = [];
+                let itemKeys = item[def.localKeys] || [];
+                itemKeys = Array.isArray(itemKeys) ? itemKeys : keys(itemKeys);
+                forEach(relatedItems, relatedItem => {
+                  if (itemKeys && contains(itemKeys, relatedItem[relationDef.idAttribute])) {
+                    attached.push(relatedItem);
+                  }
+                });
+                DSUtils.set(item, def.localField, attached);
+              });
+              return relatedItems;
+            });
+          } else if (def.type === 'belongsTo' || (def.type === 'hasOne' && def.localKey)) {
+            task = this.findAll(resourceConfig.getResource(relationName), {
+              where: {
+                [relationDef.idAttribute]: {
+                  'in': filter(map(items, item => DSUtils.get(item, def.localKey)), x => x)
+                }
+              }
+            }, __options).then(relatedItems => {
+              forEach(items, item => {
+                forEach(relatedItems, relatedItem => {
+                  if (relatedItem[relationDef.idAttribute] === item[def.localKey]) {
+                    DSUtils.set(item, def.localField, relatedItem);
+                  }
+                });
+              });
+              return relatedItems;
+            });
+          }
+
+          if (task) {
+            tasks.push(task);
+          }
+        }
+      });
+      return DSUtils.Promise.all(tasks);
+    }).then(() => items);
   }
 
   create(resourceConfig, attrs) {